started new vcfbranch instead of using master
